refactoring: moving files and making util functions
