+++ planning/autoware_lane_event_classifier/include/autoware_lane_event_classifier/detail/geometry_utils.hpp
